requirements/prod.txt
Django==4.2.9
django-betterforms==2.0.0
django-cleanup==8.1.0
django-mdeditor==0.1.20
Markdown==3.5.2
pillow==10.2.0
pymdown-extensions==10.7
python-dotenv==1.0.1
sorl-thumbnail==12.10.0